chore: example react native version upgrade
