+++ example/.node-version
@@ -1,1 +1,1 @@
-16
+18
