+++ example/app.json
@@ -1,4 +1,4 @@
 {
   "name": "SystemNavigationBarExample",
-  "displayName": "SystemNavigationBar Example"
+  "displayName": "SystemNavigationBarExample"
 }
